SelfTests: Moved ByteBuffer test to a separate project.
